# apps/api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: IUP Golf Academy API
  version: 1.0.0
  description: |
    Enterprise-grade REST API for Individual Development Plan (IUP) system for golf training academies.

    ## Features
    - Multi-tenant architecture with organization isolation
    - Role-based access control (Admin, Coach, Player)
    - JWT-based authentication with refresh tokens
    - Real-time WebSocket updates
    - Comprehensive player development tracking
    - Exercise library and session planning
    - Performance testing and benchmarking
    - Calendar and booking management
    - Analytics and reporting

    ## Authentication
    Most endpoints require authentication using JWT Bearer tokens. Obtain tokens via the `/auth/login` endpoint.

  contact:
    name: AK Golf Academy
    email: anders@akgolf.no
  license:
    name: UNLICENSED
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.iup-golf.com
    description: Production server
  - url: https://staging-api.iup-golf.com
    description: Staging server
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login
  schemas:
...
    Goal:
      type: object
      properties:
        id:
          type: string
          format: uuid
        playerId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [HANDICAP, TOURNAMENT, SKILL, FITNESS, MENTAL]
        targetValue:
          type: number
        currentValue:
          type: number
        unit:
          type: string
        targetDate:
          type: string
          format: date
        status:
          type: string
          enum: [NOT_STARTED, IN_PROGRESS, COMPLETED, ABANDONED]
        progress:
          type: number
          format: float
          minimum: 0
          maximum: 100
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
paths:
...
  /api/v1/goals:
    get:
      tags:
        - Goals
      summary: List goals
      description: Retrieve player goals (filtered by player or coach)
      parameters:
        - name: playerId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [NOT_STARTED, IN_PROGRESS, COMPLETED, ABANDONED]
      responses:
        '200':
          description: Goals retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Goal'

    post:
      tags:
        - Goals
      summary: Create goal
      description: Create new player goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - playerId
                - title
                - category
                - targetValue
                - targetDate
              properties:
                playerId:
                  type: string
                  format: uuid
                title:
                  type: string
                description:
                  type: string
                category:
                  type: string
                  enum: [HANDICAP, TOURNAMENT, SKILL, FITNESS, MENTAL]
                targetValue:
                  type: number
                unit:
                  type: string
                targetDate:
                  type: string
                  format: date
      responses:
        '201':
          description: Goal created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Goal'
  /api/v1/goals/{goalId}:
    parameters:
      - name: goalId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags:
        - Goals
      summary: Get goal by ID
      responses:
        '200':
          description: Goal retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Goal'

    patch:
      tags:
        - Goals
      summary: Update goal
      description: Update goal progress and status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                currentValue:
                  type: number
                status:
                  type: string
                  enum: [NOT_STARTED, IN_PROGRESS, COMPLETED, ABANDONED]
      responses:
        '200':
          description: Goal updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Goal'

  # ============================================================================
  # ANALYTICS ENDPOINTS
  # ============================================================================
...